Moved logic to Index from IndexManager
